split into proper architecture
