+++ docking_control.py
@@ -5,17 +5,16 @@ import math
 # eVTOL Position
 from numpy.core.umath import sign
 
-from integration_layer import turn_to_heading
+from eTaxi import get_true_position, get_imu_heading
+from navigation_control import turn_to_heading
 
-plane_x = 10.0
-plane_y = -10.0
+plane_x = 0.0
+plane_y = 0.0
 plane_heading = 0.0
 tag_size = 1 #cm
 
 # bot position, these are true values, no error
-x_pos = 0.0
-y_pos = 0.0
-heading = -math.pi/4
+
 fov_degrees = 10
 fov = math.radians(fov_degrees)
 
@@ -26,22 +25,31 @@ ERROR_MAG = 3
 X_OFFSET_MAX = 100
 
 
-def dock():
+def dock(plane_x_in, plane_y_in, plane_heading_in):
+    global plane_x, plane_y, plane_heading
+    plane_x = plane_x_in
+    plane_y = plane_y_in
+    plane_heading = plane_heading_in
+    print('plane_heading: ', math.degrees(plane_heading_in))
     turn_to_heading(plane_heading + math.pi)
     if tag_present():
-        print('tag acquired!!!!!')
+        offset = tag_x_offset()
+        print('x_offset: ', offset)
+    else:
+        print('failed to aquire tag')
 
 
 # standardized form -100 as far right, 0 is center, and 100 far left
 def tag_x_offset():
     if tag_present():
+        x_pos, y_pos = get_true_position()
         delta_x = plane_x - x_pos
         delta_y = plane_y - y_pos
         plane_angle_from_xaxis = math.atan(delta_y / delta_x)
-        angle_delta = angle_between_headings(heading, plane_angle_from_xaxis)
+        angle_delta = angle_between_headings(get_imu_heading(), plane_angle_from_xaxis)
         if angle_delta <= fov/2:
             percent_offset = angle_delta/(fov/2)
-            return percent_offset * X_OFFSET_MAX * sign(angle_delta)
+            return percent_offset * X_OFFSET_MAX
         else:
             print("something went wrong")
     else:
@@ -49,6 +57,7 @@ def tag_x_offset():
 
 
 def tag_z_distance():
+    x_pos, y_pos = get_true_position()
     delta_x = plane_x - x_pos
     delta_y = plane_y - y_pos
     dist = math.sqrt(delta_x**2 + delta_y**2)
@@ -60,6 +69,7 @@ def tag_z_distance():
 
 
 def trust_reading():
+    x_pos, y_pos = get_true_position()
     delta_x = plane_x - x_pos
     delta_y = plane_y - y_pos
     dist = math.sqrt(delta_x ** 2 + delta_y ** 2)
@@ -68,17 +78,20 @@ def trust_reading():
 
 # this needs to be fixed
 def tag_present():
+    x_pos, y_pos = get_true_position()
     delta_x = (plane_x - x_pos) + tag_size/2
     delta_y = plane_y - y_pos
     plane_angle_from_xaxis = math.atan(delta_y/delta_x)
-    return angle_between_headings(heading, plane_angle_from_xaxis) <= fov/2
+    print('plane angle from x axis: ', math.degrees(plane_angle_from_xaxis))
+    angle_diff = angle_between_headings(get_imu_heading(), plane_angle_from_xaxis)
+    print('angle_diff: ', math.degrees(angle_diff))
+    return abs(angle_diff) <= fov/2
 
 
 def angle_between_headings(angle_1, angle_2):
-    wrapped_delta = abs(angle_1- angle_2) % 360
+    wrapped_delta = abs(angle_1 - angle_2) % 2*math.pi
     shortest_delta = 2*math.pi - wrapped_delta if wrapped_delta > math.pi else wrapped_delta
-    sign = 1 if (angle_1 - angle_2 >= 0 and angle_1 - angle_2 <= 180) \
-                or (angle_1 - angle_2 <= -180 and angle_1 - angle_2 >= -360) else -1
+    sign = 1 if (angle_1 - angle_2 >= 0 and angle_1 - angle_2 <= math.pi) \
+                or (angle_1 - angle_2 <= -math.pi and angle_1 - angle_2 >= -2*math.pi) else -1
     shortest_delta *= sign
-    print('Angle Delta: ', math.degrees(shortest_delta))
-    return  shortest_delta
+    return shortest_delta
